refactor(nvml-mock): harden CUDA runtime patcher and split C out of cgo preamble
Make the in-memory interposition of the static cuda-sample safer and more
readable:

- Atomic all-or-nothing runtime patching: bounds-check every site against an
  executable /proc/self/maps mapping, save originals, and roll back if any
  write fails (a partial patch would hand host pointers to the genuine runtime
  and crash).
- Gate the amd64 attestation NOP to x86-64 (arm64 bypasses it) and guard the
  whole subsystem behind a supported-arch check so other arches compile.
- Extract the preamble C into dedicated translation units: patch.c/patch.h
  (mapping bounds check + trampoline builder) and interpose.c/interpose.h
  (segv handler, load-base finder, weak driver stubs, attestation, runtime
  interposition, export-table probe). cuda.go's preamble shrinks to #cgo +
  includes; only the four entry points Go calls are exported via headers.
- Add unit tests for the patch-eligibility logic (trampoline encoding and the
  executable-mapping check) on amd64 and arm64.

Verified on both amd64 and arm64: go test ./bridge/... passes, the c-shared
build succeeds, the weak driver stubs remain dynamically exported, and the
official vectoradd-cuda12.5.0 binary prints "Test PASSED".
